backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.8.0
python-dotenv==1.0.1
dnspython==2.7.0
ffmpeg-python==0.2.0
psutil==6.0.0
